[fix] #23 리뷰 반영 - comment를 post 하위로
