+++ src/main/java/com/abcdedu_backend/member/entity/Member.java
@@ -1,6 +1,6 @@
 package com.abcdedu_backend.member.entity;
 
-import com.abcdedu_backend.comment.Comment;
+import com.abcdedu_backend.post.comment.Comment;
 import com.abcdedu_backend.post.Post;
 import com.abcdedu_backend.utils.BaseTimeEntity;
 import jakarta.persistence.*;
