+++ src/main/java/com/abcdedu_backend/post/comment/CommentRepository.java
@@ -1,8 +1,6 @@
-package com.abcdedu_backend.comment;
+package com.abcdedu_backend.post.comment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
